backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
anthropic==0.21.7
